Move MapData.java to its own package
