+++ src/tools/image/ReliefImageBreaker.java
@@ -23,7 +23,7 @@ import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 
 import games.strategy.debug.ClientLogger;
-import games.strategy.triplea.ui.MapData;
+import games.strategy.triplea.ui.mapdata.MapData;
 import tools.map.making.ImageIoCompletionWatcher;
 import games.strategy.ui.Util;
